change task id generation     AtomicLong to boxed hashcode
add filtered callback interface
    if non-null callback is passed to create task, the result is only called back to passed callback
    if null is passed, the result is called back to all attached callback

add BLEConnectionHolder
    for singleton access

+++ peripheral_library/src/main/java/org/im97mori/ble_peripheral/BLEServerCallback.java
@@ -93,7 +93,7 @@ public interface BLEServerCallback {
      * @param value               one of {@link BluetoothGattDescriptor#ENABLE_NOTIFICATION_VALUE}, {@link BluetoothGattDescriptor#ENABLE_INDICATION_VALUE}, {@link BluetoothGattDescriptor#DISABLE_NOTIFICATION_VALUE}
      * @param argument            callback argument
      */
-    void onNotificationSuccess(long taskId, BluetoothGattServer bluetoothGattServer, BluetoothDevice device, UUID serviceUUID, UUID characteristicUUID, byte[] value, Bundle argument);
+    void onNotificationSuccess(int taskId, BluetoothGattServer bluetoothGattServer, BluetoothDevice device, UUID serviceUUID, UUID characteristicUUID, byte[] value, Bundle argument);
 
     /**
      * Notification(Indication) error callback
@@ -106,7 +106,7 @@ public interface BLEServerCallback {
      * @param status              one of {@link ErrorCodes#UNKNOWN}, {@link ErrorCodes#CANCEL}, {@link ErrorCodes#BUSY}
      * @param argument            callback argument
      */
-    void onNotificationFailed(long taskId, BluetoothGattServer bluetoothGattServer, BluetoothDevice device, UUID serviceUUID, UUID characteristicUUID, int status, Bundle argument);
+    void onNotificationFailed(int taskId, BluetoothGattServer bluetoothGattServer, BluetoothDevice device, UUID serviceUUID, UUID characteristicUUID, int status, Bundle argument);
 
     /**
      * Client Characteristic Configuration (Descriptor UUID: 0x2902) updated callback
